WarlockTmp: + Improved the uptime modeling of Metamorphosis. + Improved the modeling of cooldown collisions. + The fight length option now works. + Evolved the structure of spell classes a little more. + Added Nightfall. + Added Glyph of Corruption.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C47077

+++ Rawr.WarlockTmp/CalculationOptionsWarlock.cs
@@ -23,7 +23,7 @@ namespace Rawr.WarlockTmp {
         public String Pet { get; set; }
         public bool UseInfernal { get; set; }
         public int TargetLevel { get; set; }
-        public int Duration { get; set; }
+        public float Duration { get; set; }
         public float Latency { get; set; }
         public float ManaPotSize { get; set; }
         public float Replenishment { get; set; }
